requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.32
pyparsing==3.0.9
python-dotenv==0.21.0
python-json-logger==2.0.4
python-logstash==0.4.8
pytz==2022.6
redis==4.3.4
requests==2.28.1
six==1.16.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
